Null not being passed correctly
When a "null" was passed into SoapClient, it was being converted to a
zero. Since zero is a valid workgroup id number, the method was
returning results for workgroup zero instead of the workgroup specified
in 'group_name'. There are no workgroups with a negative id number, so
this is a simple fix.