SoapClient passing zero for null
In the definition of the workgroupGetByName method, group_id was set to
null, but the SoapClient was translating that to a zero. Zero is a
valid workgroup id and thus, incorrect results were returned. Setting
group_id to "-1" is a simple fix since there are no workgroups with
negative id numbers